fix(biens): placeholder images et compteur hero pour les biens sans photo
- Détection default.jpg + onerror fallback sur maisons/appartements/prestige/vendus
- Compteur biens intégré dans le hero avec les bonnes classes CSS
- Image Pascal Hamm ajoutée sur la page d'accueil

+++ public/admin/forgot-password.php
@@ -10,7 +10,7 @@ if (Auth::check()) {
 db()->exec("
     CREATE TABLE IF NOT EXISTS `password_reset_tokens` (
         `id`         BIGINT UNSIGNED  NOT NULL AUTO_INCREMENT,
-        `user_id`    INT UNSIGNED     NOT NULL,
+        `user_id`    BIGINT UNSIGNED  NOT NULL,
         `token_hash` VARCHAR(255)     NOT NULL,
         `expires_at` DATETIME         NOT NULL,
         `used_at`    DATETIME         DEFAULT NULL,
